+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Destiny 1 transfers are faster.
 * DIM is better at equipping exotics when you already have exotic ghosts, sparrows, and ships equipped.
 * Pulling an item from the postmaster updates the list of items quickly now.
+* Navigation from "About" or "Backers" back to your inventory works.
 
 # 4.40.0